cleaned the samples
