+++ samples/BoltOn.Samples.Application/DTOs/StudentDto.cs
@@ -4,7 +4,7 @@ namespace BoltOn.Samples.Application.DTOs
 {
 	public class StudentDto
 	{
-		public Guid Id { get; set; }
+		public Guid StudentId { get; set; }
 		public string FirstName { get; set; }
 		public string LastName { get; set; }
 		public string StudentType { get; set; }
