+++ samples/BoltOn.Samples.Console/Program.cs
@@ -1,16 +1,10 @@
 ﻿using System;
 using System.IO;
-using BoltOn.Samples.Application.Handlers;
-using Microsoft.Extensions.DependencyInjection;
-using BoltOn.Bus.MassTransit;
-using Microsoft.Extensions.Logging;
 using BoltOn.Data.EF;
-using BoltOn.Samples.Infrastructure.Data;
-using MassTransit;
-using Microsoft.EntityFrameworkCore;
+using BoltOn.Samples.Application.Handlers;
 using Microsoft.Extensions.Configuration;
-using BoltOn.Data;
-using BoltOn.Samples.Application.Entities;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Logging;
 
 namespace BoltOn.Samples.Console
 {
@@ -35,45 +29,9 @@ namespace BoltOn.Samples.Console
 			serviceCollection.BoltOn(o =>
 			{
 				o.BoltOnAssemblies(typeof(GetAllStudentsRequest).Assembly);
-				o.BoltOnMassTransitBusModule();
-				o.BoltOnCqrsModule(b => b.PurgeEventsProcessedBefore = TimeSpan.FromHours(12));
 				o.BoltOnEFModule();
 			});
 
-			var readDbConnectionString = configuration.GetValue<string>("SqlReadDbConnectionString");
-			var rabbitmqUri = configuration.GetValue<string>("RabbitMqUri");
-			var rabbitmqUsername = configuration.GetValue<string>("RabbitMqUsername");
-			var rabbitmqPassword = configuration.GetValue<string>("RabbitMqPassword");
-
-			serviceCollection.AddMassTransit(x =>
-			{
-				x.AddBus(provider => MassTransit.Bus.Factory.CreateUsingRabbitMq(cfg =>
-				{
-					cfg.Host(new Uri(rabbitmqUri), hostConfigurator =>
-					{
-						hostConfigurator.Username(rabbitmqUsername);
-						hostConfigurator.Password(rabbitmqPassword);
-					});
-
-					cfg.ReceiveEndpoint("StudentCreatedEvent_queue", ep =>
-					{
-						ep.Consumer(() => provider.GetService<BoltOnMassTransitConsumer<StudentCreatedEvent>>());
-					});
-
-					cfg.ReceiveEndpoint("StudentUpdatedEvent_queue", ep =>
-					{
-						ep.Consumer(() => provider.GetService<BoltOnMassTransitConsumer<StudentUpdatedEvent>>());
-					});
-				}));
-			});
-
-			serviceCollection.AddDbContext<SchoolReadDbContext>(options =>
-			{
-				options.UseSqlServer(readDbConnectionString);
-			});
-
-			serviceCollection.AddTransient<IRepository<StudentFlattened>, CqrsRepository<StudentFlattened, SchoolReadDbContext>>();
-			
 			var serviceProvider = serviceCollection.BuildServiceProvider();
 			serviceProvider.TightenBolts();  
 			System.Console.ReadLine();
