+++ samples/BoltOn.Samples.WebApi/PostRegistrationTask.cs
@@ -27,10 +27,6 @@ namespace BoltOn.Samples.WebApi
             writeDbContext.Set<StudentType>().Add(inState);
             writeDbContext.Set<StudentType>().Add(outOfState);
             writeDbContext.SaveChanges();
-
-			var readDbContext = scope.ServiceProvider.GetService<SchoolReadDbContext>();
-			readDbContext.Database.EnsureDeleted();
-			readDbContext.Database.EnsureCreated();
 		}
     }
 }
