+++ samples/BoltOn.Samples.WebApi/Startup.cs
@@ -37,7 +37,6 @@ namespace BoltOn.Samples.WebApi
 			services.BoltOn(options =>
 			{
 				options.BoltOnEFModule();
-				options.BoltOnMassTransitBusModule();
 				options.BoltOnCacheModule();
 				options.BoltOnWebModule();
 				options.BoltOnSerilogModule(Configuration);
@@ -46,30 +45,11 @@ namespace BoltOn.Samples.WebApi
 
 			var writeDbConnectionString = Configuration.GetValue<string>("SqlWriteDbConnectionString");
 			var readDbConnectionString = Configuration.GetValue<string>("SqlReadDbConnectionString");
-			var rabbitmqUri = Configuration.GetValue<string>("RabbitMqUri");
-			var rabbitmqUsername = Configuration.GetValue<string>("RabbitMqUsername");
-			var rabbitmqPassword = Configuration.GetValue<string>("RabbitMqPassword");
 			var redisUrl = Configuration.GetValue<string>("RedisUrl");
-			services.AddMassTransit(x =>
-			{
-				x.AddBus(provider => MassTransit.Bus.Factory.CreateUsingRabbitMq(cfg =>
-				{
-					cfg.Host(new Uri(rabbitmqUri), hostConfigurator =>
-					{
-						hostConfigurator.Username(rabbitmqUsername);
-						hostConfigurator.Password(rabbitmqPassword);
-					});
-				}));
-			});
 
 			services.AddDbContext<SchoolWriteDbContext>(options =>
 			{
 				options.UseSqlServer(writeDbConnectionString);
-			});
-
-			services.AddDbContext<SchoolReadDbContext>(options =>
-			{
-				options.UseSqlServer(readDbConnectionString);
 			});
 
 			services.AddStackExchangeRedisCache(options =>
@@ -86,7 +66,6 @@ namespace BoltOn.Samples.WebApi
 			});
 			services.AddTransient<IRepository<Student>, Repository<Student, SchoolWriteDbContext>>();
 			services.AddTransient<IQueryRepository<StudentType>, QueryRepository<StudentType, SchoolWriteDbContext>>();
-			services.AddTransient<IRepository<StudentFlattened>, Repository<StudentFlattened, SchoolReadDbContext>>();
 		}
 
 		public void Configure(IApplicationBuilder app, IHostApplicationLifetime appLifetime, IWebHostEnvironment env)
