+++ samples/docker-compose-local.yml
@@ -8,12 +8,6 @@ services:
       SA_PASSWORD: "Password1"
       ACCEPT_EULA: "Y"
     container_name: bolton-sql-container
-  rabbitmq:
-    image: rabbitmq:3.8.2-management
-    ports:
-      - "5010:5672"
-      - "5015:15672"
-    container_name: bolton-rabbitmq-container
   redis:
     image: redis
     ports:
