+++ samples/docker-compose.yml
@@ -10,14 +10,6 @@ services:
     container_name: bolton-sql-container
     networks:
       - bolton-network
-  rabbitmq:
-    image: rabbitmq:3.8.2-management
-    ports:
-      - "5010:5672"
-      - "5015:15672"
-    container_name: bolton-rabbitmq-container
-    networks:
-      - bolton-network
   redis:
     image: redis
     ports:
@@ -33,7 +25,6 @@ services:
       dockerfile: Console.Dockerfile
     container_name: bolton-console-container
     depends_on:
-      - rabbitmq
       - sql
     networks:
       - bolton-network
@@ -50,7 +41,6 @@ services:
       - "5000:80"
     container_name: bolton-api-container
     depends_on:
-      - rabbitmq
       - sql
       - console
       - redis
